chore: Added limits and checks for the standard roles

+++ cogs/settings.py
@@ -12,6 +12,8 @@ import log_helper
 
 import translation
 
+import config
+
 import database
 
 logger = log_helper.Logger("Settings")
@@ -28,6 +30,18 @@ class Settings(commands.Cog):
     ######################################################
     @stardardRolesGroup.command(name="add", description="Add a standard role to the guild")
     async def add(self, ctx, role: discord.Role):
+        # Check if the role is already a standard role
+        if database.execute_read_query(f"SELECT * FROM standard_roles WHERE guild_id={ctx.guild.id} AND role_id={role.id}"):
+            embed = discord.Embed(title=translation.get_translation(ctx.user.id, "standard_role_already_added_title"), description=translation.get_translation(ctx.user.id, "standard_role_already_added", role=role.mention), color=discord.Color.red())
+            embed.set_footer(text="Made with ❤ by the AutoVox team")
+            await ctx.response.send_message(embed=embed)
+            return
+        # Check if the role limit has been reached
+        if len(database.execute_read_query(f"SELECT * FROM standard_roles WHERE guild_id={ctx.guild.id}")) >= config.load_value("standard_roles_limit"):
+            embed = discord.Embed(title=translation.get_translation(ctx.user.id, "standard_roles_limit_reached_title"), description=translation.get_translation(ctx.user.id, "standard_roles_limit_reached", limit=config.load_value("standard_roles_limit")), color=discord.Color.red())
+            embed.set_footer(text="Made with ❤ by the AutoVox team")
+            await ctx.response.send_message(embed=embed)
+            return
         # Add a standard role to the guild
         logger.log(f"Adding standard role {role.name}({role.id}) to {ctx.guild.name}({ctx.guild.id}) by {ctx.user.name}({ctx.user.id})", log_helper.LogTypes.INFO)
         database.execute_query(f"INSERT INTO standard_roles (guild_id, role_id) VALUES ({ctx.guild.id}, {role.id})")
@@ -37,6 +51,12 @@ class Settings(commands.Cog):
 
     @stardardRolesGroup.command(name="remove", description="Remove a standard role from the guild")
     async def remove(self, ctx, role: discord.Role):
+        # Check if the role is not a standard role
+        if not database.execute_read_query(f"SELECT * FROM standard_roles WHERE guild_id={ctx.guild.id} AND role_id={role.id}"):
+            embed = discord.Embed(title=translation.get_translation(ctx.user.id, "standard_role_not_added_title"), description=translation.get_translation(ctx.user.id, "standard_role_not_added", role=role.mention), color=discord.Color.red())
+            embed.set_footer(text="Made with ❤ by the AutoVox team")
+            await ctx.response.send_message(embed=embed)
+            return
         # Remove a standard role from the guild
         logger.log(f"Removing standard role {role.name}({role.id}) from {ctx.guild.name}({ctx.guild.id}) by {ctx.user.name}({ctx.user.id})", log_helper.LogTypes.INFO)
         database.execute_query(f"DELETE FROM standard_roles WHERE guild_id={ctx.guild.id} AND role_id={role.id}")
